Fix trying to remove non-existing file
Calling rm on empty path terminated the program.

+++ hop.sh
@@ -204,8 +204,8 @@ puzzle() {
   mkdir -p "${input_path}"
   mkdir -p "${puzzle_path}"
 
-  rm "${input_path}/day${day}"
-  rm "${puzzle_path}/day${day}.md"
+  [ -f "${input_path}/day${day}" ] && rm "${input_path}/day${day}"
+  [ -f "${puzzle_path}/day${day}.md" ] && rm "${puzzle_path}/day${day}.md"
 
   aoc download --year "${year}" --day "${day}" --input-file "${input_path}/day${day}" --puzzle-file "${puzzle_path}/day${day}.md"
 }
